Fix upside-down flip

+++ shaders/sort.cs.glsl
@@ -18,7 +18,7 @@ precision highp float;
 precision highp sampler2D;
 #endif
 
-uniform uint uTileCount;
+uniform int uTileCount;
 
 layout(std430, binding = 0) buffer bTileLinkMap {
     // [0]: index of first fill in this tile
@@ -46,7 +46,7 @@ void setNext(int tileIndex, int newNextTileIndex) {
 
 void main() {
     uint globalTileIndex = gl_GlobalInvocationID.x;
-    if (globalTileIndex >= uTileCount)
+    if (globalTileIndex >= uint(uTileCount))
         return;
 
     int unsortedFirstTileIndex = getFirst(globalTileIndex);
